+++ src/main/java/kitchenpos/application/MenuService.java
@@ -1,5 +1,8 @@
 package kitchenpos.application;
 
+import kitchenpos.application.request.MenuCreateRequest;
+import kitchenpos.application.request.MenuProductCreateRequest;
+import kitchenpos.application.response.MenuResponse;
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.domain.MenuProduct;
@@ -8,9 +11,6 @@ import kitchenpos.domain.Product;
 import kitchenpos.repository.MenuGroupRepository;
 import kitchenpos.repository.MenuRepository;
 import kitchenpos.repository.ProductRepository;
-import kitchenpos.ui.request.MenuCreateRequest;
-import kitchenpos.ui.request.MenuProductCreateRequest;
-import kitchenpos.ui.response.MenuResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
