+++ src/main/java/kitchenpos/application/OrderService.java
@@ -1,5 +1,8 @@
 package kitchenpos.application;
 
+import kitchenpos.application.request.OrderCreateRequest;
+import kitchenpos.application.request.OrderLineItemCreateRequest;
+import kitchenpos.application.response.OrderResponse;
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
@@ -7,9 +10,6 @@ import kitchenpos.domain.OrderTable;
 import kitchenpos.repository.MenuRepository;
 import kitchenpos.repository.OrderRepository;
 import kitchenpos.repository.OrderTableRepository;
-import kitchenpos.ui.request.OrderCreateRequest;
-import kitchenpos.ui.request.OrderLineItemCreateRequest;
-import kitchenpos.ui.response.OrderResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
