+++ src/main/java/kitchenpos/application/TableService.java
@@ -1,13 +1,13 @@
 package kitchenpos.application;
 
+import kitchenpos.application.request.TableCreateRequest;
+import kitchenpos.application.request.TableEmptyUpdateRequest;
+import kitchenpos.application.request.TableGuestUpdateRequest;
+import kitchenpos.application.response.OrderTableResponse;
 import kitchenpos.domain.OrderStatus;
 import kitchenpos.domain.OrderTable;
 import kitchenpos.repository.OrderRepository;
 import kitchenpos.repository.OrderTableRepository;
-import kitchenpos.ui.request.TableCreateRequest;
-import kitchenpos.ui.request.TableEmptyUpdateRequest;
-import kitchenpos.ui.request.TableGuestUpdateRequest;
-import kitchenpos.ui.response.OrderTableResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
