+++ src/main/java/kitchenpos/application/request/TableCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.request;
+package kitchenpos.application.request;
 
 public class TableCreateRequest {
