+++ src/main/java/kitchenpos/application/response/TableGroupResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.response;
+package kitchenpos.application.response;
 
 import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.TableGroup;
