+++ src/main/java/kitchenpos/ui/MenuGroupRestController.java
@@ -1,8 +1,8 @@
 package kitchenpos.ui;
 
 import kitchenpos.application.MenuGroupService;
-import kitchenpos.ui.request.MenuGroupCreateRequest;
-import kitchenpos.ui.response.MenuGroupResponse;
+import kitchenpos.application.request.MenuGroupCreateRequest;
+import kitchenpos.application.response.MenuGroupResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
