+++ src/main/java/kitchenpos/ui/MenuRestController.java
@@ -1,8 +1,8 @@
 package kitchenpos.ui;
 
 import kitchenpos.application.MenuService;
-import kitchenpos.ui.request.MenuCreateRequest;
-import kitchenpos.ui.response.MenuResponse;
+import kitchenpos.application.request.MenuCreateRequest;
+import kitchenpos.application.response.MenuResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
